feat(tic-tac-toe-client): scenario tests + ADR-0066 trunk-rlib refactor (issue 430)
Phase 2 ttt-client scenarios from issue 430. Two integration tests in
`demos/tic-tac-toe-client/tests/scenario.rs`:

- `default_render_paints_grid` — load wasm, advance ticks (SDK
  auto-subscribes inputs, on_tick draws the 3x3 grid), tick before
  capture, assert `MailObserved aether.draw_triangle` and
  `DiffersFromBackground`. Validates the per-tick render path.
- `click_center_cell_drives_server_broadcast` — load both server and
  client, send synthesised `aether.window_size` + `aether.mouse_move`
  + `aether.mouse_button` to the client mailbox, advance, observe
  `tic_tac_toe.game_state` on the loopback. Validates the full chain:
  cursor caching → hit-test → `PlayMove` to server → server accepts
  + broadcasts `game_state` → bench loopback records it.

To make the click test possible (loading both server + client cdylibs
co-resident), refactor the ttt-server crate per ADR-0066. The old
`aether-demo-tic-tac-toe` was a `cdylib + rlib` crate that owned
both the wire types and the `Component` impl. When the client cdylib
linked the server's rlib for the shared `PlayMove`/`GameState`/
`MoveResult` types, the server's `#[handlers]` `aether.kinds.inputs`
section emissions stacked on top of the client's own → wasm loaded
with two Component records → substrate manifest reader rejected it
("duplicate Component record"). See issue 442 for the underlying SDK
bug; this refactor sidesteps it by adopting the camera/mesh-viewer
pattern.

Crate layout after the refactor:

- `aether-demo-tic-tac-toe` — trunk rlib. Wire types (PlayMove,
  Reset, GameState, MoveResult) + well-known mailbox names (SERVER,
  CLIENT_OBSERVER) + the GameState::new_game constructor. No
  Component impl, no FFI, no `aether-component` dep.
- `aether-demo-tic-tac-toe-server` (new) — runtime cdylib with the
  Component impl, apply_move/reply helpers, is_winner/is_board_full,
  unit tests, and `aether_component::export!`. Inherits the existing
  scenario tests (PR 437) — moved here verbatim plus the wasm
  artifact name is updated to `aether_demo_tic_tac_toe_server.wasm`.
- `aether-demo-tic-tac-toe-client` — unchanged in role; depends on
  the trunk for wire types only. The previous PR's
  `default-features = false` workaround is gone now that the trunk
  has no Component impl to gate.

CI pre-build line drops `aether-demo-tic-tac-toe` (no cdylib
anymore) and adds `aether-demo-tic-tac-toe-server` +
`aether-demo-tic-tac-toe-client` so both wasm artifacts are
available at test time. The `AETHER_REQUIRE_RUNTIME=1` env from
PR 440 keeps any forgotten entry loud.

Verified locally:
- `cargo build --target wasm32-unknown-unknown -p aether-demo-tic-tac-toe-server -p aether-demo-tic-tac-toe-client`
  succeeds in a single invocation (no two-phase build dance).
- `AETHER_REQUIRE_RUNTIME=1 cargo test -p aether-demo-tic-tac-toe -p aether-demo-tic-tac-toe-server -p aether-demo-tic-tac-toe-client`
  passes 20+ tests including 7 trunk/server unit tests, 3 server
  scenarios, 2 client scenarios, and 7 client unit tests.
- `cargo clippy --workspace --all-targets -- -D warnings` clean.
- `cargo fmt --all -- --check` clean.
- `cargo check --workspace --all-targets` clean.

Refs issue 430 (Phase 2 ttt-client checkbox + closes the
component-scenario push). Refs issue 442 (the underlying SDK
section-stacking bug this refactor sidesteps).
