+++ Cargo.toml
@@ -23,6 +23,7 @@ members = [
     "crates/aether-mail-derive",
     "demos/sokoban",
     "demos/tic-tac-toe",
+    "demos/tic-tac-toe-server",
     "demos/tic-tac-toe-client",
 ]
 # Reference artefacts under spikes/ are standalone Cargo workspaces
